Service1/requirements.txt
pika
python-json-logger
pyjwt
python-dotenv